feat: minimize right panel content re-rendering on closing/opening right panel #2450
