+++ src/dotnet/UI.Blazor.App/Components/ChatHeaderRightPanelToggle.razor
@@ -1,6 +1,6 @@
 @inherits ComputedStateComponent<bool>
 @{
-    var isVisible = !State.Value;
+    var isVisible = State.Value;
 }
 
 @if (isVisible) {
@@ -29,11 +29,16 @@
 
     protected override ComputedState<bool>.Options GetStateOptions()
         => ComputedStateComponent.GetStateOptions(GetType(),
-            static t => new ComputedState<bool>.Options() {
+            t => new ComputedState<bool>.Options() {
+                InitialValue = ScreenSize.IsNarrow() || !PanelsUI.Right.IsVisible.Value,
                 UpdateDelayer = FixedDelayer.NextTick,
                 Category = ComputedStateComponent.GetStateCategory(t),
             });
 
-    protected override Task<bool> ComputeState(CancellationToken cancellationToken)
-        => PanelsUI.Right.IsVisible.Use(cancellationToken).AsTask();
+    protected override async Task<bool> ComputeState(CancellationToken cancellationToken) {
+        if (ScreenSize.IsNarrow())
+            return true; // Do not hide button on narrow screen. Right panel will hide middle column with chat header entirely.
+
+        return !await PanelsUI.Right.IsVisible.Use(cancellationToken);
+    }
 }
